requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0
cryptography>=37.0.0
lxml>=4.9.0